requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
python-dotenv==1.0.0
fake-useragent==1.4.0
lxml==4.9.3
urllib3==2.0.7
werkzeug==2.3.7
jinja2==3.1.2
markupsafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
